Omar Tariq (aka OmarTariq612): using send() instead of sendto()
chaning sendto() function call into send() as it's being used on a tcp socket.
"If sendto() is used on a connection-mode (SOCK_STREAM,  SOCK_SEQPACKET) socket,  the arguments dest_addr and addrlen are ignored (and the error EISCONN may be returned when they are not NULL and 0)" from Linux Programmer's Manual.